refactor: Main 리팩토링
