+++ src/app/home/setting/expenseCategory/_components/DeleteExpenseCategoryModal.type.ts
@@ -1,4 +1,4 @@
-import { ExpenseCategoryAction } from "../page";
+import { ExpenseCategoryAction } from "@/reducers/expenseCategoryReducer";
 
 export interface DeleteExpenseCategoryModalProps {
     currentExpenseCategoryId: number;
