+++ src/app/home/setting/expenseCategory/_components/Main.type.ts
@@ -1,0 +1,6 @@
+export interface ExpenseCategory {
+    id: number;
+    name: string;
+    user_id: string;
+    created_at: string;
+}
